heroku-TAPI/requirements.txt
Flask
Flask-Session
tweepy
python-twitter